revert: lovelace aliases and Tip type changes
This reverts commit 445fcd0f976d771b305794cee586d825945c6af3.

Type changes made to Tip are breaking, therefore we can't
publish fix for time without reverting that change.
Type changes should go into PR #34 or into v2 devlopment branch.

+++ network.go
@@ -28,19 +28,19 @@ type (
 	// Tip defines model for tip.
 	Tip struct {
 		// Absolute Slot number (slots not divided into epochs)
-		AbsSlot uint64 `json:"abs_slot"`
+		AbsSlot int `json:"abs_slot"`
 
 		// Block Height number on chain
-		BlockNo uint64 `json:"block_no"`
+		BlockNo int `json:"block_no"`
 
 		// Timestamp for when the block was created
 		BlockTime string `json:"block_time"`
 
 		// EpochNo number
-		EpochNo uint64 `json:"epoch_no"`
+		EpochNo int `json:"epoch_no"`
 
 		// Slot number within Epoch
-		EpochSlot uint64 `json:"epoch_slot"`
+		EpochSlot int `json:"epoch_slot"`
 
 		// Block Hash in hex
 		Hash string `json:"hash"`
